Include defaults for XS functions
Many WebGPU functions have defaults in JavaScript that cannot be expressed
in C, but can be expressed in perl. Some are easy enough to express with XS.
For those, adding them in XS (and adding them to the docs) automatically
means we don't have to override them in perl. Some are more complex, and
will require perl. In those cases, we still want to try and add them to
the generated documentation.