Integrate webgpu errors with the perl exception system
As of today, wgpu-native (Mozilla's implementation) has a sizable number of
errors that as far as I can tell end up resulting in an abort instead of an
error. Things like null arguments or mistyped items. I do not currently know
the state of dawn (Google's implementation) in relation to errors