The great lcfirst-ining
Originally the thought was that keeping the names similar to the WebGPU native
version, that is with the capital letter, would be desirable. However, as time
went on, it became more and more troublesome having a different case then the
JavaScript WebGPU API. Since most examples and tutorials will use the JS API,
it became obvious that WebGPU::Direct should follow suit. It is not outside
the realm of possibility to have both available, for the time being, we are
sticking with lcfirst functions.